okf: rename enrichment_agent package to reference_agent
Reframes the OKF proof-of-concept agent as the reference agent to
disambiguate it from other agents and to match how it is described in
the README. All Python imports, ADK identifiers, CLI script, PyPI dist
name, User-Agent string, prompt filename, and prose in READMEs are
updated. SPEC.md is unchanged.
