--- okf/src/enrichment_agent/bundle/__init__.py
@@ -1,11 +1,0 @@
-from enrichment_agent.bundle.document import OKFDocument, REQUIRED_FRONTMATTER_KEYS
-from enrichment_agent.bundle.index import regenerate_indexes
-from enrichment_agent.bundle.paths import concept_id_to_path, path_to_concept_id
-
-__all__ = [
-    "OKFDocument",
-    "REQUIRED_FRONTMATTER_KEYS",
-    "concept_id_to_path",
-    "path_to_concept_id",
-    "regenerate_indexes",
-]
