--- okf/src/enrichment_agent/sources/__init__.py
@@ -1,3 +1,0 @@
-from enrichment_agent.sources.base import ConceptRef, Source
-
-__all__ = ["ConceptRef", "Source"]
