--- okf/src/enrichment_agent/viewer/__init__.py
@@ -1,3 +1,0 @@
-from enrichment_agent.viewer.generator import generate_visualization
-
-__all__ = ["generate_visualization"]
