+++ okf/src/reference_agent/__main__.py
@@ -1,4 +1,4 @@
-from enrichment_agent.cli import main
+from reference_agent.cli import main
 
 if __name__ == "__main__":
     raise SystemExit(main())
