+++ okf/src/reference_agent/agent.py
@@ -5,20 +5,20 @@ from importlib import resources
 from google.adk import Agent
 from google.adk.tools import FunctionTool
 
-from enrichment_agent.tools.bundle_tools import read_existing_doc, write_concept_doc
-from enrichment_agent.tools.source_tools import (
+from reference_agent.tools.bundle_tools import read_existing_doc, write_concept_doc
+from reference_agent.tools.source_tools import (
     list_concepts,
     read_concept_raw,
     sample_rows,
 )
-from enrichment_agent.tools.web_tools import fetch_url
+from reference_agent.tools.web_tools import fetch_url
 
 DEFAULT_MODEL = "gemini-flash-latest"
 
 
 def _load_prompt(filename: str) -> str:
     return (
-        resources.files("enrichment_agent.prompts")
+        resources.files("reference_agent.prompts")
         .joinpath(filename)
         .read_text(encoding="utf-8")
     )
@@ -26,9 +26,9 @@ def _load_prompt(filename: str) -> str:
 
 def build_bq_agent(model: str = DEFAULT_MODEL) -> Agent:
     return Agent(
-        name="okf_bq_enrichment_agent",
+        name="okf_bq_reference_agent",
         model=model,
-        instruction=_load_prompt("enrichment_instruction.md"),
+        instruction=_load_prompt("reference_instruction.md"),
         tools=[
             FunctionTool(list_concepts),
             FunctionTool(read_concept_raw),
