+++ okf/src/reference_agent/tools/web_tools.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 from typing import Any
 from urllib.parse import urlparse
 
-from enrichment_agent.tools.context import get_web_state
-from enrichment_agent.web.fetcher import FetchError, fetch_and_parse
+from reference_agent.tools.context import get_web_state
+from reference_agent.web.fetcher import FetchError, fetch_and_parse
 
 
 def fetch_url(url: str) -> dict[str, Any]:
