+++ okf/src/reference_agent/viewer/__init__.py
@@ -1,0 +1,3 @@
+from reference_agent.viewer.generator import generate_visualization
+
+__all__ = ["generate_visualization"]
